+++ lib/decorators/after.ts
@@ -1,4 +1,6 @@
-import { Request, Response, AfterCallback } from '../declarations'
+import { Request, Response, Service } from '../service'
+
+export type AfterCallback = (req: Request, res: Response, caller: Service) => Promise<void>
 
  export const After = (func: AfterCallback) => (tar: Object, _: string, descriptor: PropertyDescriptor): PropertyDescriptor => {
    const original = descriptor.value